move Context to binarylane.console package
